# api/openapi.yaml
openapi: 3.1.0
info:
  title: Subinc Cost Management Microservice API
  version: "1.0"
  description: Production-grade API for multi-tenant SaaS cloud cost management. RBAC/ABAC endpoints included.
servers:
  - url: /api
paths:
...
  /provisioning/terraform:
    post:
      summary: Enqueue a Terraform provision job
      tags: [Provisioning, Terraform]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionRequest"
      responses:
        "202":
          description: Provision job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionStatus"
    get:
      summary: List all Terraform provision jobs
      tags: [Provisioning, Terraform]
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
        - in: query
          name: org_id
          required: true
          schema:
            type: string
        - in: query
          name: project_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of provision jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProvisionStatus"
  /provisioning/terraform/{id}:
    get:
      summary: Get status of a Terraform provision job
      tags: [Provisioning, Terraform]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Provision job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /provisioning/terraform/{id}/cancel:
    post:
      summary: Cancel a Terraform provision job
      tags: [Provisioning, Terraform]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /architecture/docs:
    get:
      summary: List architecture docs
      tags: [Architecture]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of architecture docs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchitectureDoc"
    post:
      summary: Generate architecture doc
      tags: [Architecture]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDocRequest"
      responses:
        "201":
          description: Doc generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureDoc"
  /architecture/docs/{id}:
    get:
      summary: Get architecture doc
      tags: [Architecture]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Architecture doc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureDoc"
        "404":
          description: Not found
  /architecture/diagrams:
    get:
      summary: List architecture diagrams
      tags: [Architecture]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of diagrams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchitectureDiagram"
    post:
      summary: Generate architecture diagram
      tags: [Architecture]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDiagramRequest"
      responses:
        "201":
          description: Diagram generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureDiagram"
  /architecture/diagrams/{id}:
    get:
      summary: Get architecture diagram
      tags: [Architecture]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Architecture diagram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureDiagram"
        "404":
          description: Not found
...
components:
  schemas:
    ProvisionRequest:
      type: object
      required:
        - tenant_id
        - org_id
        - project_id
        - provider
        - resource
      properties:
        tenant_id:
          type: string
        org_id:
          type: string
        project_id:
          type: string
        provider:
          type: string
          description: aws, azure, gcp
        resource:
          type: string
          description: e.g., ec2, s3, vm, storage
        config:
          type: object
          additionalProperties:
            type: string
        requested_by:
          type: string
    ProvisionStatus:
      type: object
      properties:
        id:
          type: string
        request:
          $ref: "#/components/schemas/ProvisionRequest"
        status:
          type: string
          description: pending, running, success, failed, cancelled
        message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        terraform_hcl:
          type: string
          description: Terraform HCL for client export
...
    ArchitectureDoc:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        project_id:
          type: string
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        format:
          type: string
        export_url:
          type: string
        diagram_id:
          type: string
        resource_hash:
          type: string
        meta:
          type: object
          additionalProperties:
            type: string
    ArchitectureDiagram:
      type: object
      properties:
        id:
          type: string
        doc_id:
          type: string
        tenant_id:
          type: string
        project_id:
          type: string
        format:
          type: string
        created_at:
          type: string
          format: date-time
        export_url:
          type: string
        meta:
          type: object
          additionalProperties:
            type: string
    GenerateDocRequest:
      type: object
      required: [format, graph]
      properties:
        format:
          type: string
        graph:
          $ref: "#/components/schemas/ArchitectureGraph"
    GenerateDiagramRequest:
      type: object
      required: [doc_id, format, graph]
      properties:
        doc_id:
          type: string
        format:
          type: string
        graph:
          $ref: "#/components/schemas/ArchitectureGraph"
    ArchitectureGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ResourceNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/ResourceEdge"
    ResourceNode:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        provider:
          type: string
        name:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
    ResourceEdge:
      type: object
      properties:
        source_id:
          type: string
        target_id:
          type: string
        type:
          type: string
...